Improved code, separanting it in new files
