+++ src/modules/domain/validate-movement.use-case.js
@@ -1,18 +1,12 @@
 import { Player } from '../shared/constants';
 import { PiecesDataSource } from '../datasource/pieces.datasource';
 import { PlayerDataSource } from '../datasource/player.datasource';
+import { IsMovementDiagonal } from './is-movement-diagonal.use-case';
+import { IsJumpMovement } from './is-jump-movement.use-case';
 
 export const ValidateMovementUseCase = (() => {
   const execute = (origin, destination) => {
-    const possiblePiecesLocation = PlayerDataSource.getPlayerTurn() === Player.one ?
-      PiecesDataSource.getDarkPiecesPossibleLocation() : PiecesDataSource.getLightPiecesPossibleLocation();
-    if (!arrayContainsItem(possiblePiecesLocation, destination)) {
-      return false;
-    }
-
-    const currentPiecesLocation = PlayerDataSource.getPlayerTurn() === Player.one ?
-      PiecesDataSource.getDarkPiecesLocation() : PiecesDataSource.getLightPiecesLocation();  
-    if (arrayContainsItem(Object.values(currentPiecesLocation), destination)) {
+    if (!isPieceGoingToSquareThatBelongsToIt(destination) || !isPieceGoingToEmptySquare(destination)) {
       return false;
     }
 
@@ -20,40 +14,24 @@ export const ValidateMovementUseCase = (() => {
     return (isDiagonalMovement(origin, destination) && correctDirection) || isJumpMovement(origin, destination);
   }
 
-  const isDiagonalMovement = (origin, destination) => {
-    const originBoardLetter = origin[0].charCodeAt(0);
-    const originBoardNumber = origin[1];
-    const destinationBoardLetter = destination[0].charCodeAt(0);
-    const destinationBoardNumber = destination[1];
+  const isPieceGoingToSquareThatBelongsToIt = (destination) => {
+    const possiblePiecesLocation = PlayerDataSource.getPlayerTurn() === Player.one ?
+      PiecesDataSource.getDarkPiecesPossibleLocation() : PiecesDataSource.getLightPiecesPossibleLocation();
+    return arrayContainsItem(possiblePiecesLocation, destination);
+  }
 
-    return Math.abs(originBoardLetter - destinationBoardLetter) === 1 && Math.abs(originBoardNumber - destinationBoardNumber) === 1;
+  const isPieceGoingToEmptySquare = (destination) => {
+    const currentPiecesLocation = PlayerDataSource.getPlayerTurn() === Player.one ?
+      PiecesDataSource.getDarkPiecesLocation() : PiecesDataSource.getLightPiecesLocation();  
+    return !arrayContainsItem(Object.values(currentPiecesLocation), destination);
+  }
+
+  const isDiagonalMovement = (origin, destination) => {
+    return IsMovementDiagonal.execute(origin, destination);
   }
 
   const isJumpMovement = (origin, destination) => {
-    const originBoardLetter = origin[0].charCodeAt(0);
-    const originBoardNumber = +origin[1];
-    const destinationBoardLetter = destination[0].charCodeAt(0);
-    const destinationBoardNumber = +destination[1];
-
-    if (originBoardLetter === destinationBoardLetter) {
-      if (Math.abs(originBoardNumber - destinationBoardNumber) !== 2) {
-        return false
-      }
-      const locationJumped = String.fromCharCode(originBoardLetter) + ((originBoardNumber+destinationBoardNumber)/2).toString();
-      return Object.values(getOpponentPieces(PlayerDataSource.getPlayerTurn())).indexOf(locationJumped) > -1
-    } else if (originBoardNumber === destinationBoardNumber) {
-      if (Math.abs(originBoardLetter - destinationBoardLetter) !== 2) {
-        return false
-      }
-      const locationJumped = String.fromCharCode((originBoardLetter+destinationBoardLetter)/2) + originBoardNumber.toString();
-      return Object.values(getOpponentPieces(PlayerDataSource.getPlayerTurn())).indexOf(locationJumped) > -1
-    }
-
-    return false;
-  }
-
-  const getOpponentPieces = (turn) => {
-    return turn === Player.one ? PiecesDataSource.getLightPiecesLocation() : PiecesDataSource.getDarkPiecesLocation();
+    return IsJumpMovement.execute(origin, destination);
   }
 
   const arrayContainsItem = (array, item) => {
